% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTimeCols.R
\name{makeTimeCols}
\alias{makeTimeCols}
\title{Insert year, month, week columns based on an index date column}
\usage{
makeTimeCols(df, index_col, date_format = "\%Y-\%m-\%d")
}
\arguments{
\item{df}{The dataframe to modify}

\item{index_col}{Date The column with dates to index from. If character will be coerced to Date}

\item{date_format}{Character format argument to as.Date() see ?as.Date}
}
\description{
Adds year month week columns to a dataframe based on an index date column
}
