% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeValidData.R
\name{makeValidData}
\alias{makeValidData}
\title{Produce summary table of completeness and validity for variables in a dataframe}
\usage{
makeValidData(
  df,
  indexCol = "reporter",
  filtRep,
  rulesObj = NULL,
  ruleset = "default"
)
}
\arguments{
\item{df}{A dataframe whose column values are subjected to validation rules via switch()}

\item{indexCol}{The column to filter on. Should contain the values given in `filtRep`}

\item{filtRep}{Character A length 1 vector of value to filter on before}

\item{rulesObj}{List Named list of validation functions}

\item{ruleset}{Character Grouping name of custom ruleset.
calculation (e.g., input$filtRep a checkboxGroup of reporter names)}
}
\value{
Tibble with same columns as input DF, plus column "metric", and 2
rows: Completeness percentage and validity percentage
}
\description{
Runs completeness and validity calculations on input dataframe df.
Filters data by reporter to reduce process time. Use with map() to run on multiple reporters
}
\examples{
df <- data.frame('myindex' = c('Bob','Bob','Larry'),
'lastname' = c('AAA','BBB','CCC'),
'gender' = c('Male','Female','M A L E'), 'norule' = c('A','B','C'))
makeValidData(df, indexCol = 'myindex', filtRep = 'Bob')
makeValidData(df, indexCol = 'myindex', filtRep = 'Larry')
reps <- c('Bob', 'Larry')
purrr::map_df(reps, ~ makeValidData(df, indexCol = 'myindex', filtRep = .x))
}
\seealso{
Other validation: 
\code{\link{addRules}()},
\code{\link{editRules}()},
\code{\link{validRules}()},
\code{\link{validateColumn}()},
\code{\link{validateDF}()},
\code{\link{validateRows}()}
}
\concept{validation}
