% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateColumn.R
\name{validateColumn}
\alias{validateColumn}
\title{Apply validation rule from a list of rules to a column based on column name}
\usage{
validateColumn(df, col, rulesObj = NULL, ruleset = "default")
}
\arguments{
\item{df}{A dataframe whose column values are subjected to validation rules via switch()}

\item{col}{The name of the column to validate}

\item{rulesObj}{List Named list of validation functions}

\item{ruleset}{Character Grouping name of custom ruleset.}
}
\value{
Numeric value percent of column values passing validation rule from validRules()
}
\description{
Runs validation rules on specified column of a dataframe df by column name. Returns NA if no validation rule exists
}
\examples{
df <- data.frame('lastname' = c('AAA','BBB','CCC'), 'gender' = c('Male','Female','M A L E'))
validateColumn(df, 'lastname')
validateColumn(df, 'gender')

rulesObj <- validRules()
rulesObj <- rulesObj \%>\% addRules(list('norule' = function(x) x =='B'))
validateDF(df, rulesObj = rulesObj)

}
\seealso{
Other validation: 
\code{\link{addRules}()},
\code{\link{editRules}()},
\code{\link{makeValidData}()},
\code{\link{validRules}()},
\code{\link{validateDF}()}
}
\concept{validation}
