% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDF.R
\name{validateDF}
\alias{validateDF}
\title{Apply validateColumn and validRUles across a dataframe}
\usage{
validateDF(df, rulesObj = NULL, ruleset = "default")
}
\arguments{
\item{df}{A dataframe whose column values are subjected to validation rules via switch()}

\item{rulesObj}{List Named list of validation functions}

\item{ruleset}{Character Grouping name of custom ruleset.}
}
\value{
Single row dataframe with validation results and same columns as df
}
\description{
Runs validation rules on all columns of dataframe df by column name. Returns NA if no validation rule exists
}
\examples{
df <- data.frame('lastname' = c('AAA','BBB','CCC'),
'gender' = c('Male','Female','M A L E'), 'norule' = c('A','B','C'))
validateDF(df)
rulesObj <- validRules()
rulesObj <- rulesObj \%>\% addRules(list('norule' = function(x) x =='B'))
validateDF(df, rulesObj = rulesObj)

}
\seealso{
Other validation: 
\code{\link{addRules}()},
\code{\link{editRules}()},
\code{\link{makeValidData}()},
\code{\link{validRules}()},
\code{\link{validateColumn}()}
}
\concept{validation}
