% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validRules.R
\name{validRules}
\alias{validRules}
\title{Organize and access validation rules from a list of class `validRules`}
\usage{
validRules(x = list(), ruleset = "default")
}
\arguments{
\item{x}{Named List or vector of validation rule functions. Names should 
correspond to column names in a dataframe being validated}

\item{ruleset}{Character Grouping ruleset group name to assign the list in x. 
Multiple rulesets can be contained in a validRUles object. Add new 
rulesets with `addRules()`. See `?addRules`}
}
\value{
List of validation functions named for the column name they validate
}
\description{
List of validation rule functions for EDX validation use. 
THe default ruelset is from the EMSA QA2 application. If different rules 
are needed, create your own named list of rules and call validRules() to 
create the rules object. See examples.
}
\examples{
df <- data.frame("lastname" = c("AAA", "BBB"))

# Use default rules for EMSA QA2 Datasource
names(validRules())
names(validRules()[['default']])
validRules()[["default"]][["lastname"]]
validRules()[["default"]][["lastname"]](df$lastname)

# Create custom rule lists
myrules <- list(
    'columnname1' = function(x) x + 1,
    'columnname2' = function(x) x - 1
)

validRules(myrules)

myRuleSet <- validRules(myrules, ruleset = 'myRulesetName')

myrules2 <- list(
    "ruleset1" = list(
        'columnname1' = function(x) x + 1,
        'columnname2' = function(x) x - 1
    ),
    
    "ruleset2" = list(
        'columnname1' = function(x) x + 1,
        'columnname2' = function(x) x - 1
    )
)

validRules(myrules2)
validRules(myrules2) \%>\% addRules(myRuleSet, ruleset = 'myRulesetName')


}
\seealso{
Other validation: 
\code{\link{addRules}()},
\code{\link{editRules}()},
\code{\link{makeValidData}()},
\code{\link{validateColumn}()},
\code{\link{validateDF}()}
}
\concept{validation}
