% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{US Zipcodes list}
\format{
A Character vector of US 5 digit zipcodes
}
\source{
The `zipcode` column of `zipcodeDB`
https://github.com/MacHu-GWU/uszipcode-project/files/5183256/simple_db.log
}
\usage{
zipcodes
}
\description{
A vector of 5 digit US zipcodes to use for validation
}
\keyword{datasets}
